tasks/terraform.yml: Cleanup the local change, if there is
Terraform package is a bit weired and use make fmt to fix
the golang formatting locally, before building, instead
prevent the master code which is not following the golang
formatting rule with gofmt.  Because of this, during the
build process, there could be a local code change, and
leads to the go get failure.  The right fix should be
in the terraform PR process but, until that happens, we
run the task to cleanup the local code with git ansible
module to prevents the ocasional ansible run failure.